Refactor again; Keep what works
Upon testing multiple different approaches I decided upon the following:

We preserve the HTML elements the current selection includes as they are.
Trying to parse and modify them seems to volatile. The possible permutations
of various styles and HTML tags are simply too numerous. Besides we would
probably lose or at least distort the intent of the author.

Saving a text-only option should be added in addition to the existing HTML
preservation.